Adding movement to player
The movement from last program didn't work. If you went diagonally and collided with a top or bottom wall, you would go through the wall until you let go. This is because I updated the player's x and y coordinates at the same time before checking for collisions. In this version, (that works) I updated the x, checked for collisions, and then did the same for the y