Adding shootable bullets
I added bullets which are shot from the player's location, and in a straight line towards the mouse.

PROBLEM 1:
Initially, I had all the calculations in the update function, which meant that as I moved the mouse, the bullets would curve around as the calculations were done 60 times a second

SOLUTION 1:
I moved the calculations to the __init__ function instead, so that the calculation is only done once (when the bullet is created)

PROBLEM 2:
I struggled with making the bullets disappear from the walls, initially, the walls would disappear, and then the bullets would go through without doing anything.

https://gyazo.com/d1c18f15029e52697ac647724387c948