Fixed bug with player movement in new level
Previously, when the player wnet into the portal, it would move in the opposite direction in the new level. For example, if the player collided with the portal whilst going right, the player would go left automatically without the player doing anything.

To fix it, I made it so that the player has to stand still in the portal to progress.