Added shield and bullets powerup
Added shield powerup that lasts for a set amount of time.

Shield powerup:
When the shield powerup is active, the player is invincible

infinite bullets powerup:
Usually, the player is limited to only shooting 3 bullets at a time.

When you collect the infinite bullets powerup, you can shoot an infinite number of bullets.